move contrib folder
