--- contrib/gpu/docker/build.sh
@@ -1,30 +1,0 @@
-#!/bin/bash
-
-# Copyright 2022 The Pax Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-echo "Ensure you run this script from the top-level directory of the repo"
-
-CONTAINER="pax"
-if [ $# -eq 1 ]
-then
-    echo $1
-    CONTAINER=$1
-else
-    echo "Usage: bash build <container name>"
-    exit
-fi
-
-# building container here 
-docker build -t $CONTAINER . -f contrib/gpu/Dockerfile
