--- contrib/gpu/scripts_gpu/dataset_requirements.txt
@@ -1,4 +1,0 @@
-jsonlines==3.1.0
-pysimdjson==5.0.2
-jax-triton==0.1.3
-zstandard==0.18.0
