--- contrib/gpu/scripts_gpu/download_lambada.py
@@ -1,20 +1,0 @@
-# coding=utf-8
-# Copyright 2022 The Pax Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import contrib.gpu.scripts_gpu.tfds_lambada
-import tensorflow_datasets as tfds
-
-# This will download 'MyLambada' to TFDS_DATA_DIR (environment variable).
-ds = tfds.load('MyLambada')
