--- contrib/gpu/scripts_gpu/run_pile_singlenode.sh
@@ -1,31 +1,0 @@
-#! /bin/bash
-
-# coding=utf-8
-# Copyright 2022 The Pax Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-TFDS_DATA_DIR=$1
-VOCAB_PATH=$2
-LOG_DIR=${3:-"test_logdir"}
-
-export VOCAB_PATH=$VOCAB_PATH
-
-mkdir -p ${PWD}/${LOG_DIR}
-python3 /pax/paxml/paxml/main.py \
-    --job_log_dir=${PWD}/${LOG_DIR} \
-    --exp=contrib.gpu.scripts_gpu.configs.SmallPileTest \
-    --tfds_data_dir=$TFDS_DATA_DIR \
-    --should_log_compiles=True \
-    --alsologtostderr &> ${PWD}/${LOG_DIR}/output.log
-
