+++ paxml/contrib/gpu/docker/build.sh
@@ -1,4 +1,5 @@
-# coding=utf-8
+#!/bin/bash
+
 # Copyright 2022 The Pax Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,3 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+echo "Ensure you run this script from the top-level directory of the repo"
+
+CONTAINER="pax"
+if [ $# -eq 1 ]
+then
+    echo $1
+    CONTAINER=$1
+else
+    echo "Usage: bash build <container name>"
+    exit
+fi
+
+# building container here 
+docker build -t $CONTAINER . -f contrib/gpu/Dockerfile
